Remove examples feature
